UCS2113: Apply moving average filter to current measurements
As discussed in Issue #3, expected fluctuations in low-order bits
of per-port current measurements result in "flickering" power LEDs
when no downstream load is present.

To address this "noise" a moving average filter is applied to
each UCS2113 channel independently. These filters have a default
depth of 10 elements, which requires ~1s for a change to stabilize
given the 100ms interval of the main event handling loop.

Unfiltered values can be obtained from a UCS2113 instance by calling
currents() with filtered=False. The filters are still updated, but
the value directly obtained from the device is returned.